+++ CHANGELOG.md
@@ -2,6 +2,29 @@
 
 All notable changes to this project will be documented in this file. This project uses [Semantic Versioning](https://semver.org/).
 
+## 3.6.0-beta.2
+
+Special thanks to the amazing [uswong](https://github.com/uswong) for adding this new feature!
+
+### Notable Changes
+
+The new `ConfigCreator` can now be run from Homebridge UI! Once the Advantage Air device(s) is/are defined (note that only a maximum of 3 devices can be processed by `ConfigCreator`) and saved, click the `CONFIG CREATOR` button to auto-create the cmd4 configuration file required to run the `homebridge-cmd4-AdvantageAir` plugin. Make sure you check the checkbox if you want the fan to be setup as `FanSwitch` instead of a 'Fan' accessory before you click the `CONFIG CREATOR` button. 
+
+This can be run by existing users if they want to update to the newest (and best!) config options available. You can always inspect or edit the configuration created in `homebridge-cmd4` (not `homebridge-cmd4-AdvantageAir`) JASON Config editor if you want after the fact; for example if you do not like the naming convention used. Once the configuration is created, you can then use the `CHECK CONFIGURATION` button to check whether the configuration meets all the requirements. 
+
+<I><B> NOTE: </B> HOOBS users and users without access to Homebridge UI have to run the `ConfigCreator` from their terminal and follow the prompts. Instructions can be found here (instructions to come). </I>
+
+### Other Changes
+
+* Addition of `FanSwitch` config option. For users who would rather have the Advantage Air fan as a `Switch` accessory instead of a `Fan` accessory. Fan speed will still be coupled and controlled with the `Thermostat` accessory.
+* Terminal mode of `ConfigCreator.sh` now has improved instructions in coloured form.
+
+### Bug Fixes
+
+* Dev: All lint errors resolved.
+* Dev: Removed prepare script for husky
+
+
 ## 3.5.1 (2022-08-02)
 
 Special thanks to the amazing [uswong](https://github.com/uswong) for adding this new feature!
@@ -9,6 +32,7 @@ Special thanks to the amazing [uswong](https://github.com/uswong) for adding thi
 ### Bug Fixes
 
 * Replaced use of command `tac` to `sort -nr` in new `ConfigCreator` (see `v3.5.0` below). Command `tac` is not available on macOS.
+
 
 ## 3.5.0 (2022-08-02)